gen_constant_array_size_tests.py: Add tessellatino shader tests.
Test both tessellation control and tessellation evaluation shaders.

Note that this patch also adds all required extensions as explicit
requirements in the glslparsertest header; even the one in which spec
subdirectory the test resides.

XXX: Is this really neccessary? It seems highly unlikely that the compiler
ould fail to do proper constant expression in a tessellation shader stage
when the same expression worked fine in the vertex shader.